pyproject.toml
[tool.poetry]
name = "gnakrydev"
version = "0.1.3"
description = ""
authors = ["Boubacar Sidy Diallo <bsdiallo07@gmail.com>"]

[tool.poetry.scripts]
gnakrydev = "gnakrydev.main:cli"

[tool.poetry.dependencies]
python = "^3.6"
argparse = "^1.4.0"
requests = "^2.26.0"
pyaml = "^21.10.1"
uuid = "^1.30"
feedparser = "^6.0.10"
docker = "^5.0.3"
wget = "^3.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
